Move non-endpoint functions to separate modules
